requirements.txt
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.13.0
APScheduler==3.11.2
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
click-default-group==1.2.4
distro==1.9.0
docstring_parser==0.17.0
feedparser==6.0.12
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
lxml==6.0.2
openai==2.30.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
pytz==2026.1.post1
requests==2.33.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlite-fts4==1.0.3
sqlite-utils==3.39
tabulate==0.10.0
# Editable Git install with no remote (tech-content-automation==0.1.0)
-e /workspace/project
text-unidecode==1.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
